\name{amsReliability-package}
\alias{amsReliability-package}
\alias{amsReliability}
\docType{package}
\title{
\packageTitle{amsReliability}
}
\description{
\packageDescription{amsReliability}
}
\details{
\bold{CreateExportFiles()}: Use degradation and TCR files to create respective exportfiles. No arguments needed with this function.

\bold{BlackAnalysis(Scale="Lognormale",ErrorBand=TRUE,Save=TRUE)}: Using the exportfiles previously generated, perform the analysis of electromigration data. Extraction of Black parameters is performed.
}
\author{
\packageAuthor{amsReliability}

Maintainer: \packageMaintainer{amsReliability}
}
\references{

}

\keyword{ package }
\seealso{
}
\examples{

}
