% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/electromigration.r
\name{BlackAnalysis}
\alias{BlackAnalysis}
\title{Electromigration data analysis}
\usage{
BlackAnalysis(ErrorBand = TRUE, ConfidenceValue = 0.95, Save = TRUE)
}
\arguments{
\item{ErrorBand}{displays the confidence intervals if set to TRUE.}

\item{ConfidenceValue}{percentage used in the confidence interval calculation}

\item{Save}{saves the chart as .png if set to TRUE.}
}
\value{
None
}
\description{
Extract Black's parameters from a set of electromigration experiments.
The experimental data as well as the resulting model are displayed and
can be saved. Extracted parameters are saved in a fit.txt file.
}
\examples{
BlackAnalysis()
BlackAnalysis(ErrorBand=FALSE)
}
\author{
Emmanuel Chery, \email{emmanuel.chery@ams.com}
}

