% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFiles.r
\name{CreateExportFiles}
\alias{CreateExportFiles}
\title{Exportfiles creation}
\usage{
CreateExportFiles()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
Automatically creates the exportfiles for a set of electromigration experiments
by using degradation and TCR files provided by Qualitau ACE and MIRA equipments.
All files should be placed in the folder where the function is run.
The function matches file names to create the exportfiles but is able to detect
if several conditions are stored in one file. One exportfile is created for each condition.
If an exportfile is already present, user is asked if the old file is to be kept, replaced or if data have to be merged.
Times to failure are converted in seconds if they are provided in hours.
}
\examples{
CreateExportFiles()
}
\author{
Emmanuel Chery, \email{emmanuel.chery@ams.com}
}

