% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tddb.r
\name{OxideTDDB}
\alias{OxideTDDB}
\title{Oxide breakdown data analysis}
\usage{
OxideTDDB(ErrorBand = FALSE, ConfidenceValue = 0.95, Save = TRUE)
}
\arguments{
\item{ErrorBand}{displays the confidence intervals if set to TRUE.}

\item{ConfidenceValue}{percentage used in the confidence interval calculation}

\item{Save}{saves the chart as .png if set to TRUE.}
}
\value{
None
}
\description{
Extract oxide lifetime parameters from a set of Time Dependant Dielectric
Breakdown (TDDB) experiments.
The experimental data as well as the resulting model are displayed and
can be saved. Extracted parameters are saved in a fit.txt file.
}
\examples{
OxideTDDB()
OxideTDDB(ErrorBand=FALSE)
}
\author{
Emmanuel Chery, \email{emmanuel.chery@ams.com}
}

