% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.r
\name{RelAnalysis}
\alias{RelAnalysis}
\title{Reliability data analysis}
\usage{
RelAnalysis()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
This function displays a menu listing all available analyses.
User is guided with a GUI in order to select the analysis.
}
\examples{
RelAnalysis()
}
\author{
Emmanuel Chery, \email{emmanuel.chery@ams.com}
}

