\name{amsReliability-package}
\alias{amsReliability-package}
\alias{amsReliability}
\docType{package}
\title{
\packageTitle{amsReliability}
}
\description{
\packageDescription{amsReliability}
}
\details{
\bold{CreateExportFiles.Ace()}:
Use degradation and TCR files from Qualitau ACE equipments to create respective exportfiles. No arguments needed with this function.

\bold{BlackAnalysis(Scale="Lognormale",ErrorBand=TRUE,ConfidenceValue=0.95,Save=TRUE)}:
Using the exportfiles previously generated, perform the analysis of electromigration data. Extraction of Black parameters is performed. A chart is generated with the experimental data and the calculated model.
Scale can be changed to Weibull. Confidence intervals can be disabled and the confidence value can be choosen. If save is set to true, the chart is saved as PNG. 
}
\author{
\packageAuthor{amsReliability}

Maintainer: \packageMaintainer{amsReliability}
}
\references{

}

\keyword{ package }
\seealso{
}
\examples{

}
